Complete Stage 2
Present the user with a menu to choose to convert from or to a decimal number. They also have the option to exit the program.

Depending on choice, get the decimal number and base to convert to, or get the source number and source base to convert from. In either case, perform the necessary conversion and output the result.