Complete Stage 3
Upgrade the program to now handle conversion to/from any base between 2 and 36.

Re-do the menu system to now ask for the base to convert to and from, then continually get numbers until the user wants to change. (fromDecimal() and toDecimal() are now merged into convertMenu().)

Update convertDecimal() and convertNonDecimal() to use BigInteger.